refactor(csv-import): split useSetSelections into two focused hooks
useArtistSelections(sets, artistsByName) handles artist name→id mapping
across all rows. useSetMatchSelections(pageSets, pageStart, matchingSetsData)
handles create/match/duplicate defaults for the current page. Removes the
combined hook that mixed two unrelated input shapes.
